+++ Machine.h
@@ -192,7 +192,7 @@ class Machine{
 
 	int m_numberOfRanksDoneDetectingDistances;
 	// read, strand, position
-	map<int,map<char,int> > m_readsPositions;
+	map<u64,map<char,int> > m_readsPositions;
 	int m_numberOfRanksDoneSendingDistances;
 	map<int,map<int,int> > m_libraryDistances;
 	int m_libraryIterator;
